Refactor after review by Gizem and removed Crop folder
