--- src/client/components/crop/form-button/form-button.stories.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import Button from './form-button';
-
-export default {
-  title: 'Button',
-};
-export const FormButton = () => <Button placeholder="Save Crop Details" />;
