+++ src/client/components/form-field/form-field.stories.js
@@ -2,4 +2,4 @@ import React from 'react';
 import FormField from './form-field';
 
 export default { title: 'FormField' };
-export const Form = () => <FormField />;
+export const AddCropForm = () => <FormField />;
